Issue #2: Remove ES dependencies and add GP
