+++ api/jobs/index.js
@@ -1,17 +1,14 @@
 const cron = require('node-cron');
 
-const computeStats = require('./computeStats');
 const deleteInactiveUsers = require('./deleteInactiveUsers');
 const sendAlerts = require('./sendAlerts');
 
 const runJobs = () => {
   if (process.env.NODE_ENV === 'production') {
     if (process.env.PM2_INSTANCE_ID === '0') {
-      computeStats();
       deleteInactiveUsers();
     }
   } else {
-    computeStats();
     deleteInactiveUsers();
     sendAlerts();
   }
@@ -20,7 +17,6 @@ const runJobs = () => {
 const scheduleJobs = () => {
   if (process.env.NODE_ENV === 'production') {
     if (process.env.PM2_INSTANCE_ID === '0') {
-      cron.schedule('0 0 * * *', computeStats);
       cron.schedule('0 0 * * *', deleteInactiveUsers);
       cron.schedule('0 0 * * *', sendAlerts);
     }
