+++ api/package.json
@@ -10,7 +10,6 @@
   "author": "Yazan Boshmaf",
   "license": "UNLICENSED",
   "dependencies": {
-    "@elastic/elasticsearch": "^7.8.0",
     "bcryptjs": "^2.4.3",
     "colors": "^1.4.0",
     "cookie-parser": "^1.4.5",
@@ -26,6 +25,7 @@
     "mongoose": "^5.9.21",
     "node-cron": "^2.0.3",
     "nodemailer": "^6.4.10",
+    "pg": "^8.7.3",
     "xss-clean": "^0.1.1"
   },
   "devDependencies": {
