+++ docker-compose.dev.yml
@@ -41,11 +41,11 @@ services:
       RATE_LIMIT_MAX: 100
       RUN_JOBS_STARTUP: 'true'
       MONGO_URI: 'mongodb://mongo:27017/toshi'
-      ES_URI: 'http://es.cibr.qcri.org'
-      ES_PING_TIMEOUT_MS: 20000
-      ES_CRAWLER_INDEX: 'darkweb-tor-index'
-      ES_OLD_CRAWLER_SIZE: 48107548
-      ES_RECRAWLER_INDEX: 'darkweb-tor-recrawler-index'
+      GP_HOST: '10.4.8.131'
+      GP_USER: 'gpadmin'
+      GP_DATABASE: 'btc_blockchain'
+      GP_PORT: 5432
+      GP_PASSWORD: ''
       JWT_SECRET: 'UJl5Y8nK7b7TNjz8'
       JWT_AUTH_EXPIRE_DAYS: '30d'
       JWT_ACTIVATION_EXPIRE_DAYS: '3d'
